move perf@p to eval.py; fix test_data in multiprocessing
