+++ BackendBench/eval.py
@@ -205,3 +205,14 @@ def save_verbose_results(
         json.dump(results, f, indent=2)
 
     logger.info(f"Verbose results saved to {output_path}")
+
+
+def perf_at_p(correctness, performance, p=1.0):
+    assert len(correctness) == len(performance), (
+        "correctness and performance must have the same length"
+    )
+    return (
+        torch.where(torch.tensor(correctness).bool(), torch.tensor(performance) > p, 0)
+        .float()
+        .mean()
+    )
