--- BackendBench/score.py
@@ -1,18 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD 3-Clause license found in the
-# LICENSE file in the root directory of this source tree.
-
-import torch
-
-
-def perf_at_p(correctness, performance, p=1.0):
-    assert len(correctness) == len(performance), (
-        "correctness and performance must have the same length"
-    )
-    return (
-        torch.where(torch.tensor(correctness).bool(), torch.tensor(performance) > p, 0)
-        .float()
-        .mean()
-    )
